requirements.txt
grpcio>=1.56.0
grpcio-tools>=1.56.0
numpy>=1.26.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.1
requests>=2.31.0
validators>=0.22.0
pymongo>=4.3.0
protobuf==6.33.1
python-dotenv>=1.0.0
